move store to stores
