+++ project/apps/web/app/page.tsx
@@ -2,21 +2,21 @@
 
 import Link from "next/link";
 import { Button } from "@/components/ui/button";
-import { useAuthStore } from "@/store/AuthStore";
+import { useAuthStore } from "@/stores/AuthStore";
 
 const Dashboard = () => {
   const user = useAuthStore.use.user();
-  
+
   return (
     <div className="flex items-center justify-center min-h-screen">
       <div className="flex flex-col items-center">
         <div className="font-bold">Dashboard Page</div>
         <div className="flex items-center space-x-2">
-          { !user && 
+          {!user && (
             <Link href="/auth" passHref>
               <Button className="mt-4">Log In</Button>
             </Link>
-          }
+          )}
           <Link href="/questions" passHref>
             <Button className="mt-4">Go to Questions Repository</Button>
           </Link>
